requirements.txt
appdirs==1.4.4
argcomplete==1.12.3
attrs==19.3.0
cached-property==1.5.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorlog==4.8.0
defusedxml==0.6.0
distlib==0.3.2
filelock==3.0.12
google-ads==12.0.0
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.19.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleads==24.0.0
googleapis-common-protos==1.52.0
grpcio==1.38.1
httplib2==0.18.1
idna==2.10
isodate==0.6.0
lxml==4.5.2
nox==2020.12.31
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.9
pep562==1.1
proto-plus==1.18.1
protobuf==3.12.2
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pysea-common-constants==0.0.9
pysea-google-analytics==0.0.10
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.6
six==1.15.0
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.4.7
xmltodict==0.12.0
zeep==3.4.0
